RostelecomPythonProject/requirements.txt
pytest~=6.2.5
pytest-selenium
termcolor~=2.5.0
allure-python-commons

selenium~=4.9.0
python-dotenv~=1.0.1